% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coverage}
\alias{coverage}
\title{Vaccine coverage data from the WHO}
\format{
A data.table with vaccine data...
\describe{
\item{vaccine_short}{the code for for the vaccine}
\item{value}{percentage coverage from zero to 100}
...
}
}
\source{
\url{https://www.who.int/immunization/monitoring_surveillance/data/en/}
}
\usage{
coverage
}
\description{
Vaccine coverage data from the WHO
}
\keyword{datasets}
