% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demog}
\alias{demog}
\title{Demography data derived from UN World Population Prospects}
\format{
A list with three dataframes:
\describe{
\item{population}{population data for the past and future}
\item{fertility}{fertility data for the past and future}
\item{life_tables}{mx, px, and ax for the past and future}
}
}
\source{
\url{https://population.un.org/wpp/Download/Standard/Population/}
}
\usage{
demog
}
\description{
Demography data derived from UN World Population Prospects
}
\keyword{datasets}
