% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_proj.R
\name{forward_params}
\alias{forward_params}
\title{Generate the parameters for a forward step}
\usage{
forward_params(params, year)
}
\arguments{
\item{params}{A list holding all population projection parameters}

\item{year}{Integer time period}
}
\value{
A list with two elements:
(1) L - A matrix for survival and progression
(2) g - A vector with migration values
}
\description{
Generate the parameters for a forward step
}
