% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd.R
\name{gbd_rr}
\alias{gbd_rr}
\title{Calculate all-cause mortality reduction by vaccine for VIMC 10}
\usage{
gbd_rr(all_deaths, coverage_inputs)
}
\arguments{
\item{all_deaths}{All-cause deaths from calibrated pop projection}

\item{coverage_inputs}{Observed coverage}
}
\value{
A table with GBD relative-risk
}
\description{
Calculate all-cause mortality reduction by vaccine for VIMC 10
}
