% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_proj.R
\name{get_ccpm}
\alias{get_ccpm}
\title{Execute cohort-component projection model}
\usage{
get_ccpm(nx, sx, fx, mig, n)
}
\arguments{
\item{nx}{Population matrix}

\item{sx}{Survival matrix}

\item{fx}{Fertility matrix}

\item{mig}{Migration matrix}
}
\value{
A list with population and births matrices
}
\description{
Execute cohort-component projection model
}
