% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{lt_est}
\alias{lt_est}
\title{Calculate life-table values from mortality rates}
\usage{
lt_est(y)
}
\arguments{
\item{y}{Vector of mortality rates}
}
\value{
A vector of life-expectancy, 1q0, and 5q0
}
\description{
Calculate life-table values from mortality rates
}
