% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_proj.R
\name{make_params}
\alias{make_params}
\title{Generate population projection parameters from WPP data}
\usage{
make_params(wpp_dt)
}
\arguments{
\item{wpp_dt}{A data.table with WPP data}
}
\value{
A list of parameters
}
\description{
Generate population projection parameters from WPP data
}
