% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{map_locations}
\alias{map_locations}
\title{Make a map showing presence or absence of an indicator}
\usage{
map_locations(locations, title)
}
\arguments{
\item{locations}{A character vector of iso3 codes for locations}

\item{title}{A string with the title of the plot}
}
\value{
A ggplot object with a world map
}
\description{
Make a map showing presence or absence of an indicator
}
\examples{
map_locations(loc_table$location_iso3, "All locations")
}
