% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.coverage}
\alias{plot.coverage}
\title{Plot coverage for a specific location}
\usage{
\method{plot}{coverage}(x, ...)
}
\arguments{
\item{dt}{A data.table with coverage data}
}
\value{
A plot of the coverage data
}
\description{
Plot coverage for a specific location
}
