% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_coverage.R
\name{prep_coverage}
\alias{prep_coverage}
\title{Pull all vaccine coverage data for user supplied list of locations}
\usage{
prep_coverage(iso3 = NULL)
}
\arguments{
\item{iso3}{Vector of iso3 codes}
}
\value{
A data.table with historical coverage for all vaccines for each
supplied location
}
\description{
Pull all vaccine coverage data for user supplied list of locations
}
