% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_coverage.R
\name{prep_hpv_coverage_data}
\alias{prep_hpv_coverage_data}
\title{Pull in HPV coverage data}
\usage{
prep_hpv_coverage_data()
}
\value{
A data.table with all HPV data
}
\description{
Pull in HPV coverage data
}
