% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_coverage.R
\name{prep_reported_coverage_data}
\alias{prep_reported_coverage_data}
\title{Pull in reported coverage data and bind it together}
\usage{
prep_reported_coverage_data()
}
\value{
A data.table with reported coverage data for Japanese encephalitis
and Meningitis A
}
\description{
Pull in reported coverage data and bind it together
}
