% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_demog.R
\name{prep_wpp_data}
\alias{prep_wpp_data}
\title{Pull in relevant datasets from WPP 2019}
\usage{
prep_wpp_data(
  datasets = c("tfr", "popM", "popF", "mxF", "mxM", "migration", "sexRatio",
    "percentASFR")
)
}
\arguments{
\item{datasets}{A character vector with names of valid datasets in the
\code{wpp2019} package}
}
\value{
A data.table with each of the requested datasets bound together
}
\description{
Pull in relevant datasets from WPP 2019
}
\examples{
prep_wpp_data(c("popM", "popF"))
}
