% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_coverage.R
\name{prep_wuenic_data}
\alias{prep_wuenic_data}
\title{Pull in WUENIC data and bind it together}
\usage{
prep_wuenic_data()
}
\value{
A data.table with all WUENIC data
}
\description{
Pull in WUENIC data and bind it together
}
