% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_proj.R
\name{project_pop}
\alias{project_pop}
\title{Project the population}
\usage{
project_pop(params)
}
\arguments{
\item{params}{A list holding all population projection parameters}
}
\value{
A matrix with the population in all time periods
}
\description{
Project the population
}
