% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_proj.R
\name{step_forward}
\alias{step_forward}
\title{Step the population forward one time period}
\usage{
step_forward(n, params, year)
}
\arguments{
\item{n}{A vector with the current population}

\item{params}{A list holding all population projection parameters}

\item{year}{Integer time period}
}
\value{
A vector with the population one time period forward
}
\description{
Step the population forward one time period
}
