% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_coverage.R
\name{total_coverage2}
\alias{total_coverage2}
\title{Calculate total coverage for every vaccine, location, sex}
\usage{
total_coverage2(coverage)
}
\description{
Calculate total coverage for every vaccine, location, sex
}
