% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vimc_impact}
\alias{vimc_impact}
\title{VIMC vaccine impact estimates}
\format{
A data.table with vaccine impacts estimates from VIMC...
\describe{
\item{vaccine_short}{the code for for the vaccine}
\item{value}{deaths averted from the vaccine}
...
}
}
\source{
\url{https://montagu.vaccineimpact.org/}
}
\usage{
vimc_impact
}
\description{
VIMC vaccine impact estimates
}
\keyword{datasets}
