% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimc_mort_reduct.R
\name{vimc_mort_reduct}
\alias{vimc_mort_reduct}
\title{Calculate all-cause mortality reduction by vaccine for VIMC 10}
\usage{
vimc_mort_reduct(wpp_input, obs_wpp, vimc)
}
\arguments{
\item{wpp_input}{Input WPP data}

\item{obs_wpp}{Observed WPP data}

\item{vimc}{VIMC deaths averted}
}
\value{
A list of tables with population projection results
}
\description{
Calculate all-cause mortality reduction by vaccine for VIMC 10
}
