% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimc.R
\name{vimc_rr}
\alias{vimc_rr}
\title{Calculate all-cause mortality reduction by vaccine for VIMC 10}
\usage{
vimc_rr(wpp_input, vimc_impact_estimates)
}
\arguments{
\item{wpp_input}{Input WPP data}

\item{vimc_impact_estimates}{VIMC deaths averted}
}
\value{
A list of tables with population projection results
}
\description{
Calculate all-cause mortality reduction by vaccine for VIMC 10
}
