% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd.R
\name{gbd_rr}
\alias{gbd_rr}
\title{Calculate all-cause mortality reduction by vaccine for GBD}
\usage{
gbd_rr(wpp_input, obs_wpp)
}
\arguments{
\item{wpp_input}{Input WPP data}

\item{obs_wpp}{Observed WPP data}
}
\value{
A list of tables with population projection results
}
\description{
Calculate all-cause mortality reduction by vaccine for GBD
}
