% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{get_ccpm}
\alias{get_ccpm}
\title{Execute cohort-component projection model}
\usage{
get_ccpm(nx, sx, fx, mig, z, n)
}
\arguments{
\item{nx}{Population matrix}

\item{sx}{Survival matrix}

\item{fx}{Fertility matrix}

\item{mig}{Migration matrix}

\item{z}{Number of age groups}

\item{n}{Number of years of projection}
}
\value{
A list with population and births matrices
}
\description{
Execute cohort-component projection model
}
