% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_proj.R
\name{project_pop}
\alias{project_pop}
\title{Run the population projection for a particular country and year range}
\usage{
project_pop(is, y0, y1, wpp_input, obs_wpp)
}
\arguments{
\item{is}{Location name}

\item{y0}{Start year of projection}

\item{y1}{End year of projection}

\item{wpp_input}{Input WPP data}

\item{obs_wpp}{Observed WPP data}
}
\value{
A list of tables with population projection results
}
\description{
Run the population projection for a particular country and year range
}
